Задание 7.1 выведены картинки
