+++ js/main.js
@@ -1,18 +1,1 @@
-import { getRandom, checkLength } from './utils.js';
-import { createAuthor, createMock, createAuthors } from './data.js';
-import { } from './loading.js';
-
-const post = {
-  id: 1,
-  url: "",
-  description: "",
-  likes: 15,
-  comments: [],
-};
-
-const comment = {
-  id: 135,
-  avatar: 'img/avatar-6.svg',
-  message: 'В целом всё неплохо. Но не всё.',
-  name: 'Артём',
-};
+import './loading.js';
